requirement.txt
Django>=4.2
djangorestframework
django-cors-headers
django-environ
drf-yasg
celery>=5.0
mysqlclient
